Added file src/sim/sim_neurospaceselement.c for creating a genesis compartment via neurospaces.
So when executing PidinLookupTopSymbol it returns NULL which is an error.

Source of error could possibly be PidinStackParse does not properly create
the associative data structure for the lookup. 

Also need to look into creating a persistent data struct for neurospaces elements
to be loaded when genesis starts and stays in memory until execution is finished.

Will rename the file to sim_neurospacescompartment.c later.