Tests are currently broken dude to a change made in sim_create for looking up a parent.
creating an object like /h/c/n would not work due to the failure for 
genesis to look up /h/c since that object resides in the neurospaces 
model container. A proper lookup for /h/c must be hooked into the create 
function when getting the parent element.

Added the olf libarary from gen 2.3. Some of the unecessary code from it 
must still be removed. 

More work must be done on nscreate's tabchannel creation functionality 
to properly set the heiarchy for the objects.