Added the necessary code for creating a channel in nscreate.
Added code for allowing channel fields to be set in nssetfield. It 
parses correctly however when looking at fields via print info it 
appears that the values are not being set correctly. A sentinel value is 
in place which is set by Heccer.

Added a test for nesting several compartments and channels.