Fixed the issue of assigning the proper parameter to the according children has been fixed.
Name lookup had an issue. Must remember if we're doing a lookup on a 
symbol that we push via a PidinStackPushString that we need to make a 
copy of it.

Added some default set parameters to the ChannelCalloc and the 
HHGate Calloc operations to prevent heccer from producing errors. 

Next is to work on the setupalpha function which will set up the rest of 
the parameters for the channels.