Support for tabchannels in place.
They are added in certain functions when certain conditions in the script are met.
  in: 
    nssetfield.c - to create activation, inactivation and concentration gates when seeing X,Y or Z 
    nstabcreate.c - to create parameterized channels
    nssetup.c - to add parameters to channel objects.

tests are in tables.t

table-kc is a tabulated channel and thus far does not evaluate correctly, it expects parameters
and does not use the tabulated values. This could be due to setting flags wrong causing the heccer
logic to think it's parameterized instead of tabulated.