ASC_FILE: Added the parsing support for the ascii output for the heccer object.
TESTS: Removed all of the instances where the parser automatically sets 
state_init values for gates. Now they are all set via X_init, Y_init, 
and Z_init with the test scripts updated accordingly. 

channel2-nernst1: It doesn't look like this test will have output that 
will line up with the heccer output without some hacky work. The script 
detects the two nernst instances, one in each compartment and produces 
output from it's function. This output is not present int he heccer 
expected output, so filtering of output must be done on the seen output 
rather than the expected. Even so, some of the numerical values are a 
bit off, most likely due to the presence of the nernst function 
calculations on the reversal potential.