BUGFIX: Fixed the neurospaces_harness' ability to perform a test off of files. It now properly deletes generated application files and wirtes /tmp/text_seen and /tmp/text_expected like the other tests.
Added ActMsg for use in adding an activation message to a synchan. There is a problem 
however since the heccer instances are created unpon the "reset" command, and the 
message setup it always enountered BEFORE a reset. Will need to make an intermediate 
data structure for holding the act message data ahead of time and have it used during 
the reset process.