Added a new test for copying a non rooted source before proceeding.
Added a commodity function for obtaining a rooted context. There is an issue with NSplit, as it is set to some arbitrary number and needs to be parsed and set correctly. "x" "y" and "z" parameters are not properly looked up. 

Getting a malloc incorrect checksum error when performing a pidinstack parse on /test/primary_dend when it is parent to one of the second_dend objects. Also getting a double definition of some second_dend objects when they haven't been created yet.