Removed a bug which causes len to become infinity, this then causes NSplit and nlambda to become infinity causing a seg fault.
Added a proper prototype to parse_compartment.

There is still an issue with free being called on some memory that has not been allocated or has had its address changed. It seems to occur in 
PidinStackLookupTopSymbol after allocating the second context. Attempted to free the first context to perform the second lookup safely but to no 
avail. Will investigate further.