Rearranged the tests specifications so that the output is more comprehensible.
The command html-upload-prepare now works. I needed to install HTML::Tables and HTML::Template to get it to work. 

Readcell now outputs the number of compartments read in during a parse. 

Todo: Next we must allow more heccer objects to be constructed by keeping a list of objects to be created in the heccer array. nsintegrator will most likely keep a dynami array of strings with the neutral object names to be given a heccer object and they will be created when calling NSReset. Similar to the IO Messages.