Found the issue with the globals not being transferred to the model container. GENESIS will crea te the variables ELEAK and EREST_ACT if they are not already explicitly declared and will leave them empty.
Added the purkinje cell model as well as defaults.g and vclamp.g to the manifest. 

Need to add support for the diffamp object which is used to make the Vclamp used for the simulation. Added the dev_struct.h, diffamp.c and devlib.g to the neurospaces directory to start work on adding in the object.